+++ Changelog.md
@@ -1,84 +1,112 @@
-# v1.0.0
+# v1.1.0
+
 ## Changes
+
+1. Cleaning code
+2. Switch to ReVanced Extended patches
+3. Switch to curl instead wget
+4. Restore YouTube Music Root options
+
+## Fixes
+
+1. Improve stability of script
+
+# v1.0.0
+
+## Changes
+
 1. Drop YouTube Music Root building
 2. Move to python
 
 # v0.9.3
+
 ## Changes
+
 1. YouTube update to 17.49.37
 2. YouTube Music update to 5.36.51
 3. Fix revanced integrations download
 
 # v0.9.2
+
 ## Changes
+
 1. YouTube update to 17.31.35
 2. YouTube Music update to 5.17.51
 
 # v0.9.1
+
 ## Changes
+
 1. Improve adb connection
 2. Add adb checking
 
 # v0.9
+
 ## Changes
+
 1. All script now running in powershell
 2. Add downloading Java. Make script for releases much lighter
 3. Check Execution Policy
 
 ## Fixes
+
 1. Downloading now fixed
 
 # v0.8
+
 ## Changes
+
 1. Return options to not download revanced resource
 2. Add downloading YT or YT Music. Its increase speed of publishing releases much faster
 
 # v0.7.1
+
 ## Changes
+
 1. Now code looks like much better
 2. Message that appear more clearer
 
 # v0.7
+
 ## Changes
+
 1. Update YT apk to 17.29.34
 2. Script rewrite. Now you input ADB device number at first time and choose what do you want to build
 3. Excluded custom-branding patches
 4. Add message to close jar builiding window
 
-# v0.6.1 
+# v0.6.1
+
 ## Changes
+
 1. Update YT apk to 17.28.34
 2. Exclude premium-heading and sponsorblock patches. Include swipe-controls
 
 # v0.6
+
 ## Changes
+
 1. Add menu that propose install stock apk
 2. Add menu that propose uninstall stock apk
 
 ## Fixes
+
 1. ReVanced team was fix something so now you can build YouTube Music Root
 
 # v0.5.1
+
 ## Changes
+
 1. Add checking apk files in folder
 
 # v0.5
+
 ## Changes
+
 1. Finally add Non-Root version to build
 2. Update YouTube 17.27.39 and YouTube Music 5.14.53 apk
 3. Add to release mircog.apk for Non-Root version. You need to install it manually
 
 ## Fixes
+
 1. Script delete unnecessary .dex file
-
-
-
-
-
-
-
-
-
-
-
-